# openapi-schema.yml
openapi: 3.0.2
info:
  title: ''
  version: ''
paths:
...
  /devices/{id}/history/:
    get:
      operationId: DeviceHistoryDevice
      description: ''
      parameters:
      - name: id
        in: path
        required: true
        description: A unique integer value identifying this device.
        schema:
          type: string
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Device'
          description: ''
      tags:
      - devices
...
  /devices/{id}/allocate/:
    patch:
      operationId: deviceAllocateDevice
      description: ''
      parameters:
      - name: id
        in: path
        required: true
        description: A unique integer value identifying this device.
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Device'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/Device'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/Device'
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Device'
          description: ''
      tags:
      - devices
components:
  schemas:
...
    Device:
      type: object
      properties:
        id:
          type: integer
          readOnly: true
        deviceName:
          type: string
          maxLength: 100
        deviceUser:
          type: string
          readOnly: true
        deviceType:
          type: integer
      required:
      - deviceName
      - deviceType
...